fix: change DistanceAngleConfig fields from int to double
The approaching_closed_doors.distance config value is 5.2 (float),
which caused JSON deserialization failures with int-typed properties.

+++ src/SPTQuestingBots.Server/Configuration/QuestingBotsConfig.cs
@@ -141,10 +141,10 @@ public class MinMaxConfig
 public class DistanceAngleConfig
 {
     [JsonProperty("distance")]
-    public int Distance { get; set; }
+    public double Distance { get; set; }
 
     [JsonProperty("angle")]
-    public int Angle { get; set; }
+    public double Angle { get; set; }
 }
 
 // ────────────────────────────────────────────────────────────────────────
